[simulation_kernel] fix jira 3313 : "L'affichage des vitesses des unités dans gaming est mal arrondi"

+++ src/libraries/simulation_kernel/Adapters/RolePion_Location.cpp
@@ -209,7 +209,7 @@ namespace
         }
         virtual void SendFullState( client::UnitAttributes& message )
         {
-            message().set_speed( static_cast< unsigned int >( MIL_Tools::ConvertSpeedSimToMos( *speed_ ) ) );
+            message().set_speed( MIL_Tools::ConvertSpeedSimToMos( *speed_ ) );
         }
     private:
         void Changed()
